Implement JWT tokens

+++ Controllers/BaseController.cs
@@ -2,7 +2,10 @@ using System.Reflection;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using Caker.Models;
+using Caker.Models.Interfaces;
 using Caker.Repositories;
+using Caker.Services.CurrentUserService;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 
@@ -11,25 +14,57 @@ namespace Caker.Controllers
     [ApiController]
     [Route("api/[controller]")]
     public abstract class BaseController<TModel, TDto, TCreateDto, TUpdateDto>(
-        BaseRepository<TModel> repository
+        BaseRepository<TModel> repository,
+        ICurrentUserService currentUserService
     ) : ControllerBase
-        where TModel : BaseModel
+        where TModel : BaseModel, IDtoable<TDto>, IAccessibleBy
         where TCreateDto : class
         where TUpdateDto : class
     {
         protected readonly BaseRepository<TModel> _repository = repository;
+        protected readonly ICurrentUserService _currentUserService = currentUserService;
+
+        protected bool CanCreate(TModel? model)
+        {
+            try
+            {
+                _currentUserService.GetUserId();
+                return true;
+            }
+            catch
+            {
+                return false;
+            }
+        }
+
+        protected bool CanRead(TModel? model) => true;
+
+        protected bool CanUpdate(TModel? model) => CheckPermissions(model);
+
+        protected bool CanDelete(TModel? model) => CheckPermissions(model);
 
         protected abstract TModel CreateModel(TCreateDto dto);
         protected abstract void UpdateModel(TModel model, TUpdateDto dto);
-        protected abstract TDto ToDto(TModel model);
+
+        protected virtual bool CheckPermissions(TModel? model)
+        {
+            if (!_currentUserService.IsAuthorized())
+                return false;
+
+            return _currentUserService.IsAdmin()
+                || (model?.AllowedUserIds.Contains(_currentUserService.GetUserId()) ?? true);
+        }
 
         [HttpGet]
         public virtual async Task<ActionResult<IEnumerable<TDto>>> GetAll()
         {
+            if (!CanRead(null))
+                return Forbid();
+
             try
             {
                 var result = await _repository.GetAll();
-                return Ok(result.Select(ToDto));
+                return Ok(result.Select(m => m.ToDto()));
             }
             catch (Exception ex)
             {
@@ -41,17 +76,26 @@ namespace Caker.Controllers
         public virtual async Task<ActionResult<TDto>> GetById(int id)
         {
             var entity = await _repository.GetById(id);
-            return entity == null ? NotFound() : Ok(ToDto(entity));
+
+            if (!CanRead(entity))
+                return Forbid();
+
+            return entity == null ? NotFound() : Ok(entity.ToDto());
         }
 
+        [Authorize]
         [HttpPost]
         public virtual async Task<ActionResult<TDto>> Create([FromBody] TCreateDto dto)
         {
             try
             {
                 var entity = CreateModel(dto);
+
+                if (!CanCreate(entity))
+                    return Forbid();
+
                 await _repository.Create(entity);
-                return CreatedAtAction(nameof(GetById), new { id = entity.Id }, ToDto(entity));
+                return CreatedAtAction(nameof(GetById), new { id = entity.Id }, entity.ToDto());
             }
             catch (DbUpdateException ex)
             {
@@ -60,19 +104,22 @@ namespace Caker.Controllers
         }
 
         [HttpPut("{id}")]
-        public virtual async Task<IActionResult> Update(int id, [FromBody] TUpdateDto dto)
+        public virtual async Task<ActionResult<TDto>> Update(int id, [FromBody] TUpdateDto dto)
         {
             var entity = await _repository.GetById(id);
             if (entity == null)
                 return NotFound();
 
+            if (!CanUpdate(entity))
+                return Forbid();
+
             UpdateModel(entity, dto);
             await _repository.Update(entity);
-            return NoContent();
+            return Ok(entity.ToDto());
         }
 
         [HttpPatch("{id}")]
-        public virtual async Task<IActionResult> PartialUpdate(
+        public virtual async Task<ActionResult<TDto>> PartialUpdate(
             int id,
             [FromBody] JsonElement patchDoc
         )
@@ -80,6 +127,9 @@ namespace Caker.Controllers
             var entity = await _repository.GetById(id);
             if (entity == null)
                 return NotFound();
+
+            if (!CanUpdate(entity))
+                return Forbid();
 
             var modelType = typeof(TModel);
             var properties = modelType.GetProperties(BindingFlags.Public | BindingFlags.Instance);
@@ -115,12 +165,15 @@ namespace Caker.Controllers
             }
 
             await _repository.Update(entity);
-            return NoContent();
+            return Ok(entity.ToDto());
         }
 
         [HttpDelete("{id}")]
         public virtual async Task<IActionResult> Delete(int id)
         {
+            if (!CanDelete(await _repository.GetById(id)))
+                return Forbid();
+
             await _repository.Delete(id);
             return NoContent();
         }
